refactor(apa-lesson): modularize validators
