+++ packages/domain/src/apa-lesson/services/index.ts
@@ -1,1 +1,9 @@
-export * from "./validation";
+export * from "./active-recall-phase-validator";
+export * from "./assimilation-phase-validator";
+export * from "./feedback-next-phase-validator";
+export * from "./lesson-validation-service";
+export * from "./lesson-validator";
+export * from "./linguistic-objective-validator";
+export * from "./phase-base-validator";
+export * from "./presentation-phase-validator";
+export * from "./validation-types";
