+++ packages/domain/src/apa-lesson/services/validation-types.ts
@@ -1,0 +1,19 @@
+export type ValidationSeverity = "error" | "warning";
+
+export interface ValidationIssue {
+  path: string;
+  message: string;
+  severity: ValidationSeverity;
+}
+
+export interface ValidationResult {
+  valid: boolean;
+  issues: ValidationIssue[];
+}
+
+export function getValidationStatus(issues: ValidationIssue[]): ValidationResult {
+  return {
+    valid: !issues.some((issue) => issue.severity === "error"),
+    issues,
+  };
+}
